Add files via upload
I've uploaded another version of my Healthcare analysis, but this time, it was done using Microsoft Excel. All results and sights are inside a PowerPoint presentation. The reason for my excel analysis is to have a version focusing on a specific section and have knowledge of that section. In this case, we covered various topics that are affected by gender.